- move all resultproxy tests intio test_resultset, adapted from 03d35a833cff68013781640b1192d079ab7e1154 in master (not cherry-picked)
